backend_orchestrator/pyproject.toml
[tool.poetry]
name = "backend-orchestrator"
version = "1.0.0"
description = "Central orchestration service for healthcare chatbot AI agents"
authors = ["Prawibowo <tan.joenarto@bithealth.co.id>"]
readme = "README.md"
packages = [{include = "src"}]

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "0.104.1"
uvicorn = {extras = ["standard"], version = "0.24.0"}
pydantic = "2.5.0"
pydantic-settings = "2.1.0"
redis = "5.0.1"
confluent-kafka = "2.3.0"
psycopg2-binary = "2.9.9"
sqlalchemy = "2.0.23"
alembic = "1.13.1"
google-generativeai = "0.3.2"
anthropic = "0.7.8"
sentence-transformers = "2.2.2"
numpy = ">=1.26.0"
python-multipart = "0.0.6"
python-jose = {extras = ["cryptography"], version = "3.3.0"}
passlib = {extras = ["bcrypt"], version = "1.7.4"}
httpx = "0.25.2"
pyyaml = "6.0.1"
structlog = "23.2.0"
prometheus-client = "0.19.0"

[tool.poetry.group.dev.dependencies]
pytest = "7.4.3"
pytest-asyncio = "0.21.1"
pytest-mock = "3.12.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
